replaced shadcn alert with new custom alert dialog
