+++ apps/web/components/shared/AlertDialog.tsx
@@ -1,150 +1,45 @@
 "use client";
 
-import * as React from "react";
-import * as AlertDialogPrimitive from "@radix-ui/react-alert-dialog";
+import Modal from "@/components/shared/Modal";
+import { Button } from "@formbricks/ui";
 
-import { cn } from "@formbricks/lib/cn";
-
-const AlertDialog: React.ComponentType<AlertDialogPrimitive.AlertDialogContentProps> =
-  AlertDialogPrimitive.Root;
-
-const AlertDialogTrigger: React.ComponentType<AlertDialogPrimitive.AlertDialogTriggerProps> =
-  AlertDialogPrimitive.Trigger;
-
-const AlertDialogPortal = ({
-  className,
-  children,
-  ...props
-}: AlertDialogPrimitive.AlertDialogPortalProps) => (
-  <AlertDialogPrimitive.Portal className={cn(className)} {...props}>
-    <div className="fixed inset-0 z-50 flex items-end justify-center sm:items-center">{children}</div>
-  </AlertDialogPrimitive.Portal>
-);
-AlertDialogPortal.displayName = AlertDialogPrimitive.Portal.displayName;
-
-const AlertDialogOverlay = React.forwardRef<
-  React.ElementRef<typeof AlertDialogPrimitive.Overlay>,
-  React.ComponentPropsWithoutRef<typeof AlertDialogPrimitive.Overlay>
->(({ className, children, ...props }, ref) => (
-  <AlertDialogPrimitive.Overlay
-    className={cn(
-      "animate-in fade-in fixed inset-0 z-50 bg-black/50 backdrop-blur-sm transition-opacity",
-      className
-    )}
-    {...props}
-    ref={ref}
-  />
-));
-AlertDialogOverlay.displayName = AlertDialogPrimitive.Overlay.displayName;
-
-const AlertDialogContent: React.ComponentType<AlertDialogPrimitive.AlertDialogContentProps> =
-  React.forwardRef<
-    React.ElementRef<typeof AlertDialogPrimitive.Content>,
-    React.ComponentPropsWithoutRef<typeof AlertDialogPrimitive.Content>
-  >(({ className, ...props }, ref) => (
-    <AlertDialogPortal>
-      <AlertDialogOverlay />
-      <AlertDialogPrimitive.Content
-        ref={ref}
-        className={cn(
-          "animate-in fade-in-90 slide-in-from-bottom-10 sm:zoom-in-90 sm:slide-in-from-bottom-0 fixed z-50 grid w-full max-w-lg scale-100 gap-4 bg-white p-6 opacity-100 sm:rounded-lg md:w-full",
-          "dark:bg-slate-900",
-          className
-        )}
-        {...props}
-      />
-    </AlertDialogPortal>
-  ));
-AlertDialogContent.displayName = AlertDialogPrimitive.Content.displayName;
-
-interface AlertDialogHeaderProps extends React.HTMLAttributes<HTMLDivElement> {
-  dangerouslySetInnerHTML?: {
-    __html: string;
-  };
+interface AlertDialogProps {
+  open: boolean;
+  setOpen: (open: boolean) => void;
+  confirmWhat: string;
+  onDiscard: () => void;
+  text?: string;
+  useSaveInsteadOfCancel?: boolean;
+  onSave?: () => void;
 }
 
-const AlertDialogHeader = ({ className, ...props }: AlertDialogHeaderProps) => (
-  <div className={cn("flex flex-col space-y-2 text-center sm:text-left", className)} {...props} />
-);
-AlertDialogHeader.displayName = "AlertDialogHeader";
-
-interface AlertDialogFooterProps extends React.HTMLAttributes<HTMLDivElement> {
-  dangerouslySetInnerHTML?: {
-    __html: string;
-  };
+export default function AlertDialog({
+  open,
+  setOpen,
+  confirmWhat,
+  onDiscard,
+  text,
+  useSaveInsteadOfCancel = false,
+  onSave,
+}: AlertDialogProps) {
+  return (
+    <Modal open={open} setOpen={setOpen} title={`Confirm ${confirmWhat}`}>
+      <p>{text || "Are you sure? This action cannot be undone."}</p>
+      <div className="my-4 space-x-2 text-right">
+        <Button
+          variant="secondary"
+          onClick={() => {
+            if (useSaveInsteadOfCancel && onSave) {
+              onSave();
+            }
+            setOpen(false);
+          }}>
+          {useSaveInsteadOfCancel ? "Save" : "Cancel"}
+        </Button>
+        <Button variant="warn" onClick={onDiscard}>
+          Discard
+        </Button>
+      </div>
+    </Modal>
+  );
 }
-
-const AlertDialogFooter = ({ className, ...props }: AlertDialogFooterProps) => (
-  <div
-    className={cn("flex flex-col-reverse sm:flex-row sm:justify-end sm:space-x-2", className)}
-    {...props}
-  />
-);
-AlertDialogFooter.displayName = "AlertDialogFooter";
-
-const AlertDialogTitle: React.ComponentType<AlertDialogPrimitive.AlertDialogTitleProps> = React.forwardRef<
-  React.ElementRef<typeof AlertDialogPrimitive.Title>,
-  React.ComponentPropsWithoutRef<typeof AlertDialogPrimitive.Title>
->(({ className, ...props }, ref) => (
-  <AlertDialogPrimitive.Title
-    ref={ref}
-    className={cn("text-lg font-semibold text-slate-900", "dark:text-slate-50", className)}
-    {...props}
-  />
-));
-AlertDialogTitle.displayName = AlertDialogPrimitive.Title.displayName;
-
-const AlertDialogDescription: React.ComponentType<AlertDialogPrimitive.AlertDialogDescriptionProps> =
-  React.forwardRef<
-    React.ElementRef<typeof AlertDialogPrimitive.Description>,
-    React.ComponentPropsWithoutRef<typeof AlertDialogPrimitive.Description>
-  >(({ className, ...props }, ref) => (
-    <AlertDialogPrimitive.Description
-      ref={ref}
-      className={cn("text-sm text-slate-500", "dark:text-slate-400", className)}
-      {...props}
-    />
-  ));
-AlertDialogDescription.displayName = AlertDialogPrimitive.Description.displayName;
-
-const AlertDialogAction: React.ComponentType<AlertDialogPrimitive.AlertDialogActionProps> = React.forwardRef<
-  React.ElementRef<typeof AlertDialogPrimitive.Action>,
-  React.ComponentPropsWithoutRef<typeof AlertDialogPrimitive.Action>
->(({ className, ...props }, ref) => (
-  <AlertDialogPrimitive.Action
-    ref={ref}
-    className={cn(
-      "inline-flex h-10 items-center justify-center rounded-md bg-slate-900 px-4 py-2 text-sm font-semibold text-white transition-colors hover:bg-slate-700 focus:outline-none focus:ring-2 focus:ring-slate-400 focus:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50 dark:bg-slate-100 dark:text-slate-900 dark:hover:bg-slate-200 dark:focus:ring-slate-400 dark:focus:ring-offset-slate-900",
-      className
-    )}
-    {...props}
-  />
-));
-AlertDialogAction.displayName = AlertDialogPrimitive.Action.displayName;
-
-const AlertDialogCancel: React.ComponentType<AlertDialogPrimitive.AlertDialogCancelProps> = React.forwardRef<
-  React.ElementRef<typeof AlertDialogPrimitive.Cancel>,
-  React.ComponentPropsWithoutRef<typeof AlertDialogPrimitive.Cancel>
->(({ className, ...props }, ref) => (
-  <AlertDialogPrimitive.Cancel
-    ref={ref}
-    className={cn(
-      "mt-2 inline-flex h-10 items-center justify-center rounded-md border border-slate-200 bg-transparent px-4 py-2 text-sm font-semibold text-slate-900 transition-colors hover:bg-slate-100 focus:outline-none focus:ring-2 focus:ring-slate-400 focus:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50 dark:border-slate-700 dark:text-slate-100 dark:hover:bg-slate-700 dark:focus:ring-slate-400 dark:focus:ring-offset-slate-900 sm:mt-0",
-      className
-    )}
-    {...props}
-  />
-));
-AlertDialogCancel.displayName = AlertDialogPrimitive.Cancel.displayName;
-
-export {
-  AlertDialog,
-  AlertDialogTrigger,
-  AlertDialogContent,
-  AlertDialogHeader,
-  AlertDialogFooter,
-  AlertDialogTitle,
-  AlertDialogDescription,
-  AlertDialogAction,
-  AlertDialogCancel,
-};
